BUGGED - Classes

+++ Person.cpp
@@ -4,12 +4,12 @@ Person::Person (string name, string email):
 	mName(name), mEmail(email)
 {}
 
-Student::Student(string name, string email, int gpa) : Person(name, email), mGpa(gpa)
+Student::Student(string name, string email, float gpa) : Person(name, email), mGpa(gpa)
 {
 
 }
 
-Teacher::Teacher (string name, string email, int salary):Person(name, email), mSalary(salary)
+Teacher::Teacher (string name, string email, float salary):Person(name, email), mSalary(salary)
 {
 
 }
